==> Notebooks/01_get_tv_shows_df.ipynb <==
import requests
import bs4
import pandas as pd
import numpy as np

## Get TV shows from each streaming service

#### Goal

- Use webscraping to get the tv shows
- Export each data frame as csv

#### General information:

- Netflix has 1,915 tv shows, 50 per page, 39 pages in total.

- Prime has 2,136 tv shows, 50 per page, 43 in total.
- HBO has 200 tv shows, 50 per page, 4 pages in total.


### 1. Define function to get tv shows and additional information

We will use webscarping to get the tv shows from the website reelgood.com.

Each page has up to 50 tv shows.

# Define function that takes the url and number of tv shows you want to get from it, default is 50 which is the maximum.

def get_tv_shows(url, no=50):
    
    # define empty variables
    shows = {}
    year = []
    rating = []
    imdb = []
    rotten_tom = []
    
    # get response and contents using beautiful soup
    resp = requests.get(url)
    soup = bs4.BeautifulSoup(resp.content, "html.parser")
    tv_titles = soup.find("table", class_="css-1179hly").find_all("td", class_="css-1u7zfla")
    extras = soup.find("table", class_="css-1179hly").find_all("td", class_="css-1u7zfla")
    
    # get list of tv shows
    shows["show"] = [tv_titles[i].find("a").contents[0] for i in range(no)]
    
    # get lists for years, rating, imdb and rotten tomatoes reviews
    for i in range(no):
        year.append(soup.find("table", class_="css-1179hly").find_all("td", class_="css-1u11l3y")[i*4+0].contents[0])
        rating.append(soup.find("table", class_="css-1179hly").find_all("td", class_="css-1u11l3y")[i*4+1].contents[0])
        imdb.append(soup.find("table", class_="css-1179hly").find_all("td", class_="css-1u11l3y")[i*4+2].contents[0])
        rotten_tom.append(soup.find("table", class_="css-1179hly").find_all("td", class_="css-1u11l3y")[i*4+3].contents[0])

    shows["year"] = year
    shows["rating"] = rating
    shows["imdb"] = imdb
    shows["rotten_tomatoes"] = rotten_tom
    
    return shows

## 2. Get tv shows for each streaming service using get_tv_shows function

Next, we will create a while loop in order to run the function up until the number of tv shows in each and updating the url to change pages and get the next 50 tv shows. We can run this loop until the last number divisible by 50, do for the last page that has less than 50 tv shows, we will run an individual code cell and add it to the data frame.

Afterwards, each data frame will be exported to csv.

## Netflix

url = "https://reelgood.com/tv/source/netflix"
pages = 0
netflix_shows = pd.DataFrame(columns=["show", "year", "rating", "imdb", "rotten_tomatoes"])

while pages < 1900:
    new_pd = pd.DataFrame(get_tv_shows(url))
    netflix_shows = netflix_shows.append(new_pd)
    pages +=50
    url = "https://reelgood.com/tv/source/netflix?offset=" + str(pages)

netflix_shows

netflix_shows = netflix_shows.append(pd.DataFrame(get_tv_shows("https://reelgood.com/tv/source/netflix?offset=1900", 15)))


netflix_shows = netflix_shows.reset_index(drop=True)

netflix_shows

# netflix_shows.to_csv(("../Data/netflix_shows.csv"), index_label=False)

## Amazon Prime

url = "https://reelgood.com/tv/source/amazon"
pages = 0
amazon_shows = pd.DataFrame(columns=["show", "year", "rating", "imdb", "rotten_tomatoes"])

while pages < 2100:
    new_pd = pd.DataFrame(get_tv_shows(url))
    amazon_shows = amazon_shows.append(new_pd)
    pages +=50
    url = "https://reelgood.com/tv/source/amazon?offset=" + str(pages)

amazon_shows = amazon_shows.append(pd.DataFrame(get_tv_shows("https://reelgood.com/tv/source/amazon?offset=2100", 36)))


amazon_shows = amazon_shows.reset_index(drop=True)

amazon_shows

# amazon_shows.to_csv(("../Data/amazon_shows.csv"), index_label=False)

## HBO

url = "https://reelgood.com/tv/source/hbo"
pages = 0
hbo_shows = pd.DataFrame(columns=["show", "year", "rating", "imdb", "rotten_tomatoes"])

while pages < 200:
    new_pd = pd.DataFrame(get_tv_shows(url))
    hbo_shows = hbo_shows.append(new_pd)
    pages +=50
    url = "https://reelgood.com/tv/source/hbo?offset=" + str(pages)

hbo_shows = hbo_shows.reset_index(drop=True)

hbo_shows

# hbo_shows.to_csv(("../Data/hbo_shows.csv"), index_label=False)
