==> Notebooks/02_imdb_tv_shows.ipynb <==
import pandas as pd
import numpy as np
pd.set_option("display.max_columns", None)

## Get imdb tv shows data frame

#### Goals

- Import imdb title basics csv

- Filter only tv show related titles

- Export final data frame

titles = pd.read_csv("C:/Users/Admin/Documents/ironhack/title.basics.tsv.gz", sep="\t", low_memory=False)

titles.head()

titles["titleType"].value_counts()

tv_all = titles.loc[np.where(titles["titleType"].isin(["tvSeries", "tvMiniSeries", "tvShort", "tvSpecial"]))].reset_index(drop=True)


tv_all.dtypes

**We will change the startYear and endYear columns to numeric columns before exporting the data frame to a pickle, in order to keep the data types.**

tv_all["startYear"] = pd.to_numeric(tv_all["startYear"], errors="coerce")

tv_all["endYear"] = pd.to_numeric(tv_all["startYear"], errors="coerce")

# tv_all.to_pickle("../Data/imdb_tv_all.pkl")
