==> Notebooks/03.1_a_netflix_imdb_merge.ipynb <==
import pandas as pd
import numpy as np

## Netflix Merge

#### Goals:

- Merge Netflix tv shows with IMDB ID


**NOTE: Since we need to do different mergings in order to get the complete data related for each streaming service, we will do it in individual notebooks for each.**


## 1. Import imdb and netflix data

imdb = pd.read_pickle("../Data/imdb_tv_all.pkl")

netflix = pd.read_csv("../Data/netflix_shows.csv")

We will look at how many titles we have using the shape function, which will return a tuple indicating number of rows and number of columns.

netflix.shape

## 2. Merge dataframes

We will start by merging on title and year in order to get the right IMDB ID (tconst) because there might be some remakes.
We will do this by using a left join in order to keep all the original netflix titles.

netflix_genres = netflix.merge(imdb, left_on=["show", "year"], right_on=["originalTitle", "startYear"], how="left")

From the new merged data frame, we will get a new data frame called netflix_missing1 to get the missing titles that didn't find a match.

netflix_missing1 = netflix_genres[netflix_genres["tconst"].isna()].reset_index(drop=True)[["show", "year", "rating"
                                                                                           , "imdb", "rotten_tomatoes"]]

netflix_missing1.shape

We are missing 649 titles from 1915 in total.

We will now merge but now using primaryTitle instead of originalTitle, since some of the titles might be in their original language.

netflix_missing1 = netflix_missing1.merge(imdb, left_on=["show", "year"], right_on=["primaryTitle", "startYear"], how="left")

We will create another data frame for those titles who weren't matched.

netflix_missing2 = netflix_missing1[netflix_missing1["tconst"].isna()].reset_index(drop=True)[["show", "year", "rating"
                                                                                               , "imdb", "rotten_tomatoes"]]

netflix_missing2.shape

We are now missing 452 titles.

We will merge again without the year, since there might be some discrepancy on the years taken from web scraping.

netflix_missing2 = netflix_missing2.merge(imdb, left_on=["show"], right_on=["originalTitle"], how="left")

We will create a data frame for missing titles each time we merge.

netflix_missing3 = netflix_missing2[netflix_missing2["tconst"].isna()].reset_index(drop=True)[["show", "year", "rating"
                                                                                               , "imdb", "rotten_tomatoes"]]

netflix_missing3.shape

We are now missing 324 titles. We will now try using the primaryTitle.

netflix_missing3 = netflix_missing3.merge(imdb, left_on=["show"], right_on=["primaryTitle"], how="left")

netflix_missing4 = netflix_missing3[netflix_missing3["tconst"].isna()].reset_index(drop=True)[["show", "year", "rating"
                                                                                               , "imdb", "rotten_tomatoes"]]

netflix_missing4.shape

We are now missing 311 titles. Since we don't have something else to relate to in this dataframe, we will now import the other data frame from the IMDB data base which includes all different titles created for a single one and the IMDB ID (titleID).

title_regions = pd.read_csv("C:/Users/Admin/Documents/ironhack/title.akas.tsv.gz", sep="\t", low_memory=False)

title_regions.head()

We will do the merge of the missing titles with the new title_regions data frame by title name.

netflix_missing4 = netflix_missing4.merge(title_regions, left_on=["show"], right_on=["title"], how="left")

netflix_missing5 = netflix_missing4[netflix_missing4["titleId"].isna()].reset_index(drop=True)[["show", "year", "rating"
                                                                                                , "imdb", "rotten_tomatoes"]]

netflix_missing5.shape

Now we are just missing 198 titles, we will export this data frame and work on it on the next notebook.

## 3. Export data frames

We will join all data frames without null values to get a final netflix data frame containing the IMDB ID.
We will drop duplicates because the title regions may have joined multiple rows since some titles have the same name in different regions.

netflix_genres = netflix_genres[~netflix_genres["tconst"].isna()]

netflix_genres = netflix_genres.append(netflix_missing1[~netflix_missing1["tconst"].isna()])

netflix_genres = netflix_genres.append(netflix_missing2[~netflix_missing2["tconst"].isna()]).drop_duplicates("show")

netflix_genres = netflix_genres.append(netflix_missing3[~netflix_missing3["tconst"].isna()]).drop_duplicates("show")

netflix_genres = netflix_genres.append(netflix_missing4[~netflix_missing4["titleId"].isna()]).drop_duplicates("show")

netflix_genres.columns

netflix_genres = netflix_genres.reset_index(drop=True)

netflix_genres.head()

# netflix_genres.to_pickle("../Data/netflix_ids.pkl")

We will also drop duplicates from the missing titles data frame.

netflix_missing5 = netflix_missing5.drop_duplicates("show")

netflix_missing5.head()

# netflix_missing5.to_pickle("../Data/netflix_missing.pkl")
