==> Notebooks/03.1_b_amazon_imdb_merge.ipynb <==
import pandas as pd
import numpy as np

## Amazon Merge

#### Goals:

- Merge Amazon tv shows with IMDB ID

We will follow the same steps as in 03_a_netflix_imdb_merge notebook.

## 1. Import imdb and amazon data

imdb = pd.read_pickle("../Data/imdb_tv_all.pkl")

amazon = pd.read_csv("../Data/amazon_shows.csv")

amazon.shape

We have 2136 titles from the original data frame.

## 2. Merge dataframes

We will start by merging on title and year in order to get the right IMDB ID (tconst) because there might be some remakes.
We will do this by using a left join in order to keep all the original amazon titles.

amazon_genres = amazon.merge(imdb, left_on=["show", "year"], right_on=["originalTitle", "startYear"], how="left")

From the new merged data frame, we will get a new data frame called hbo_missing1 to get the missing titles that didn't find a match. We will do this in each step.

amazon_missing1 = amazon_genres[amazon_genres["tconst"].isna()].reset_index(drop=True)[["show", "year", "rating", "imdb"
                                                                                        , "rotten_tomatoes"]]

amazon_missing1.shape

We are missing 1033 titles from 2136 in total.

We will now merge but now using primaryTitle instead of originalTitle, since some of the titles might be in their original language.

amazon_missing1 = amazon_missing1.merge(imdb, left_on=["show", "year"], right_on=["primaryTitle", "startYear"], how="left")

amazon_missing2 = amazon_missing1[amazon_missing1["tconst"].isna()].reset_index(drop=True)[["show", "year", "rating"
                                                                                            , "imdb", "rotten_tomatoes"]]

amazon_missing2.shape

We are now missing 968 titles.

We will merge again without the year, since there might be some discrepancy on the years taken from web scraping.

amazon_missing2 = amazon_missing2.merge(imdb, left_on=["show"], right_on=["originalTitle"], how="left")

amazon_missing3 = amazon_missing2[amazon_missing2["tconst"].isna()].reset_index(drop=True)[["show", "year", "rating"
                                                                                            , "imdb", "rotten_tomatoes"]]

amazon_missing3.shape

We are now missing 724 titles. We will now try using the primaryTitle.

amazon_missing3 = amazon_missing3.merge(imdb, left_on=["show"], right_on=["primaryTitle"], how="left")

amazon_missing4 = amazon_missing3[amazon_missing3["tconst"].isna()].reset_index(drop=True)[["show", "year", "rating"
                                                                                            , "imdb", "rotten_tomatoes"]]

amazon_missing4.shape

We are now missing 311 titles. Since we don't have something else to relate to in this dataframe, we will now import the other data frame from the IMDB data base which includes all different titles created for a single one and the IMDB ID (titleID).

title_regions = pd.read_csv("C:/Users/Admin/Documents/ironhack/title.akas.tsv.gz", sep="\t", low_memory=False)

title_regions.head()

We will do the merge of the missing titles with the new title_regions data frame by title name.

amazon_missing4 = amazon_missing4.merge(title_regions, left_on=["show"], right_on=["title"], how="left")

amazon_missing5 = amazon_missing4[amazon_missing4["titleId"].isna()].reset_index(drop=True)[["show", "year", "rating"
                                                                                             , "imdb", "rotten_tomatoes"]]

amazon_missing5.shape

Now we are just missing 596 titles, we will export this data frame and work on it on the next notebook.

## 3. Export data frames

We will join all data frames without null values to get a final amazon data frame containing the IMDB ID.
We will drop duplicates because the title regions may have joined multiple rows since some titles have the same name in different regions.

amazon_genres = amazon_genres[~amazon_genres["tconst"].isna()]

amazon_genres = amazon_genres.append(amazon_missing1[~amazon_missing1["tconst"].isna()])

amazon_genres = amazon_genres.append(amazon_missing2[~amazon_missing2["tconst"].isna()]).drop_duplicates("show")

amazon_genres = amazon_genres.append(amazon_missing3[~amazon_missing3["tconst"].isna()]).drop_duplicates("show")

amazon_genres = amazon_genres.append(amazon_missing4[~amazon_missing4["titleId"].isna()]).drop_duplicates("show")

amazon_genres.columns

amazon_genres = amazon_genres.reset_index(drop=True)

amazon_genres.head()

# amazon_genres.to_pickle("../Data/amazon_ids.pkl")

We will also drop duplicates from the missing titles data frame.

amazon_missing5 = amazon_missing5.drop_duplicates("show")

amazon_missing5.head()

# amazon_missing5.to_pickle("../Data/amazon_missing.pkl")
