==> Notebooks/03.1_c_hbo_merge.ipynb <==
import pandas as pd
import numpy as np

## HBO Merge

#### Goals:

- Merge HBO tv shows with IMDB ID

We will follow the same steps as in 03_a_netflix_imdb_merge notebook.

## 1. Import imdb and hbo data

imdb = pd.read_pickle("../Data/imdb_tv_all.pkl")

hbo = pd.read_csv("../Data/hbo_shows.csv")

hbo.shape

We have 200 titles from the original data frame.

## 2. Merge dataframes

We will start by merging on title and year in order to get the right IMDB ID (tconst) because there might be some remakes.
We will do this by using a left join in order to keep all the original hbo titles.

hbo_genres = hbo.merge(imdb, left_on=["show", "year"], right_on=["originalTitle", "startYear"], how="left")

From the new merged data frame, we will get a new data frame called hbo_missing1 to get the missing titles that didn't find a match. We will do this in each step.

hbo_missing1 = hbo_genres[hbo_genres["tconst"].isna()].reset_index(drop=True)[["show", "year", "rating", "imdb"
                                                                               , "rotten_tomatoes"]]

hbo_missing1.shape

We are missing 41 titles from 200 in total.

We will now merge but now using primaryTitle instead of originalTitle, since some of the titles might be in their original language.

hbo_missing1 = hbo_missing1.merge(imdb, left_on=["show", "year"], right_on=["primaryTitle", "startYear"], how="left")

hbo_missing2 = hbo_missing1[hbo_missing1["tconst"].isna()].reset_index(drop=True)[["show", "year", "rating", "imdb"
                                                                                   , "rotten_tomatoes"]]

hbo_missing2.shape

We are now missing 34 titles.

We will merge again without the year, since there might be some discrepancy on the years taken from web scraping.

hbo_missing2 = hbo_missing2.merge(imdb, left_on=["show"], right_on=["originalTitle"], how="left")

hbo_missing3 = hbo_missing2[hbo_missing2["tconst"].isna()].reset_index(drop=True)[["show", "year", "rating", "imdb"
                                                                                   , "rotten_tomatoes"]]

hbo_missing3.shape

We are now missing 20 titles. We will now try using the primaryTitle.

hbo_missing3 = hbo_missing3.merge(imdb, left_on=["show"], right_on=["primaryTitle"], how="left")

hbo_missing4 = hbo_missing3[hbo_missing3["tconst"].isna()].reset_index(drop=True)[["show", "year", "rating", "imdb"
                                                                                   , "rotten_tomatoes"]]

hbo_missing4.shape

We are now missing 19 titles. Since we don't have something else to relate to in this dataframe, we will now import the other data frame from the IMDB data base which includes all different titles created for a single one and the IMDB ID (titleID).

title_regions = pd.read_csv("C:/Users/Admin/Documents/ironhack/title.akas.tsv.gz", sep="\t", low_memory=False)

title_regions.head()

We will do the merge of the missing titles with the new title_regions data frame by title name.

hbo_missing4 = hbo_missing4.merge(title_regions, left_on=["show"], right_on=["title"], how="left")

hbo_missing5 = hbo_missing4[hbo_missing4["titleId"].isna()].reset_index(drop=True)[["show", "year", "rating"
                                                                                             , "imdb", "rotten_tomatoes"]]

hbo_missing5.shape

Now we are just missing 15 titles, we will export this data frame and work on it on the next notebook.

## 3. Export data frames

We will join all data frames without null values to get a final hbo data frame containing the IMDB ID.
We will drop duplicates because the title regions may have joined multiple rows since some titles have the same name in different regions.

hbo_genres = hbo_genres[~hbo_genres["tconst"].isna()]

hbo_genres = hbo_genres.append(hbo_missing1[~hbo_missing1["tconst"].isna()])

hbo_genres = hbo_genres.append(hbo_missing2[~hbo_missing2["tconst"].isna()]).drop_duplicates("show")

hbo_genres = hbo_genres.append(hbo_missing3[~hbo_missing3["tconst"].isna()]).drop_duplicates("show")

hbo_genres = hbo_genres.append(hbo_missing4[~hbo_missing4["titleId"].isna()]).drop_duplicates("show")

hbo_genres.columns

hbo_genres = hbo_genres.reset_index(drop=True)

hbo_genres.head()

# hbo_genres.to_pickle("../Data/hbo_ids.pkl")

We will also drop duplicates from the missing titles data frame.

hbo_missing5 = hbo_missing5.drop_duplicates("show")

hbo_missing5 = hbo_missing5.reset_index(drop=True)

hbo_missing5.head()

# hbo_missing5.to_pickle("../Data/hbo_missing.pkl")
