==> Notebooks/03.2_a_netflix_missing_title_ids.ipynb <==
import pandas as pd
import numpy as np
from fuzzywuzzy import fuzz

## Find missing IMDB IDs for Netflix Data

#### Goals

- Find missing titles from original merge.
- Find missing titles on IMDB using fuzzywuzzy.

### 1. Import data

netflix = pd.read_csv("../Data/netflix_shows.csv")

netflix_ids = pd.read_pickle("../Data/netflix_ids.pkl")

netflix_missing = pd.read_pickle("../Data/netflix_missing.pkl")

### 2. Compare length of values of oringal list vs new ones

We will compare the number of rows from the original netflix data frame and the sum of the missing titles and the netflix ids.

len(netflix)

len(netflix_ids) + len(netflix_missing)

netflix["show"].nunique()

6 values are missing but when looking at unique values on the original data frame, we get the same result. This means that there must be some repeated titles which may be from different years. We will make a value count for the show column in the original netflix data frame.

netflix["show"].value_counts().head(6)

We will look at each of them and add them the netflix missing data frame.

netflix[netflix["show"] == "The Code"]

netflix_ids[netflix_ids["show"] == "The Code"]

netflix_missing = netflix_missing.append(netflix[(netflix["show"] == "The Code") & (netflix["year"] == 2014)])

netflix[netflix["show"] == "Maniac"]

netflix_ids[netflix_ids["show"] == "Maniac"]

netflix_missing = netflix_missing.append(netflix[(netflix["show"] == "Maniac") & (netflix["year"] == 2015)])

netflix[netflix["show"] == "Danger Mouse"]

netflix_ids[netflix_ids["show"] == "Danger Mouse"]

netflix_missing = netflix_missing.append(netflix[(netflix["show"] == "Danger Mouse") & (netflix["year"] == 2015)])

netflix[netflix["show"] == "Charmed"]

netflix_ids[netflix_ids["show"] == "Charmed"]

netflix_missing = netflix_missing.append(netflix[(netflix["show"] == "Charmed") & (netflix["year"] == 2018)])

netflix[netflix["show"] == "The Good Cop"]

netflix_ids[netflix_ids["show"] == "The Good Cop"]

netflix_missing = netflix_missing.append(netflix[(netflix["show"] == "The Good Cop") & (netflix["year"] == 2015)])

netflix[netflix["show"] == "Inspector Gadget"]

netflix_ids[netflix_ids["show"] == "Inspector Gadget"]

netflix_missing = netflix_missing.append(netflix[(netflix["show"] == "Inspector Gadget") & (netflix["year"] == 2015)])

len(netflix_missing) + len(netflix_ids)

Now we got the same values from the original tv shows and the new data frames.

netflix_missing = netflix_missing.reset_index(drop=True)

### 3. Find missing titles using fuzzywuzzy

We will first create a list for missing shows.

netflix_shows_missing = [show for show in netflix_missing["show"]]

len(netflix_shows_missing)

imdb = pd.read_pickle(path + "Data/imdb_tv_all.pkl")

We will create a function to find the most similar title on the imdb title basics file.

imdb_titles = [title for title in imdb["primaryTitle"]]

def find_shows(show):
    matches = []

    for title in imdb_titles:
        # compute ratio and remove case-sensitivity
        ratio = fuzz.ratio(title.lower(), show.lower())

        # add all matches to list with ratio > 60
        if ratio >= 60:
            matches.append((title, show, ratio))
    
    # return none if there was no match found
    if len(matches) == 0:
        return None
    return sorted(matches, key=lambda x: x[2], reverse=True)[0][0]


We will make a test with the first value from the missing shows list.

netflix_shows_missing[0]

find_shows(netflix_shows_missing[0])

We will now create a new column with the matched imdb titles.

netflix_missing["imdb_titles"] = netflix_missing["show"].apply(lambda x: find_shows(x))

netflix_missing.head(10)

From the first 10 missing titles we can see that our function did a good job, we will now take a look at the titles that weren't found a match.

netflix_missing[netflix_missing["imdb_titles"].isna()]

netflix_missing = netflix_missing[~netflix_missing["imdb_titles"].isna()]

Since we have only 4 titles and they are in another language, we will remove them from our data.

### 4. Merge IMDB IDs

We will still merge on the year, in order to be sure that we are not merging a remake or a similar title.

netflix_missing = netflix_missing.merge(imdb, how="left", left_on=["imdb_titles", "year"], right_on=["primaryTitle", "startYear"])

netflix_missing.head()

We will now take a look at the titles who didn't find a match.

netflix_missing[netflix_missing["tconst"].isna()]

We can see that most of this titles don't have a similar imdb titles, due to this, we will drop this columns also.

netflix_ids_new = netflix_missing[~netflix_missing["tconst"].isna()].reset_index(drop=True).drop_duplicates("tconst")

netflix_ids = netflix_ids.append(netflix_ids_new)

netflix_ids.head()

We will now calculate the ratio of titles with an imdb id match and the original titles.

len(netflix_ids) / len(netflix)

This means that we will be using 95% of the original data.

### 5. Cleaning final data

We will create a final data frame including:
- show
- year
- rating
- imdb
- rotten_tomatoes
- imdb_id
- all data from title basics

First, we will need to create the imdb_id column, this will have tconst and titleId values.

netflix_ids["imdb_id"] = np.where(netflix_ids["tconst"].isna(), netflix_ids["titleId"], netflix_ids["tconst"])

netflix_ids["imdb_id"].isna().value_counts()

This means we have all the IMDB IDs for all the titles in this final data frame. We will now remove columns that are not show, year, rating, imdb, rotten_tomatoes and imdb_id and merge again.

to_drop = [col for col in netflix_ids.columns if col not in ["show", "year", "rating", "imdb", "rotten_tomatoes", "imdb_id"]]

netflix_ids = netflix_ids.drop(columns=to_drop)

netflix_ids.head()

Now we will merge all data from title basics

netflix_ids = netflix_ids.merge(imdb, how="left", left_on="imdb_id", right_on="tconst")

netflix_ids.head()

netflix_ids = netflix_ids.drop(columns="tconst")

### 6. Export final data

# netflix_ids.to_pickle("../Data/netflix_final.pkl")
