==> Notebooks/03.2_b_amazon_missing_title_ids.ipynb <==
import pandas as pd
import numpy as np
from fuzzywuzzy import fuzz

## Find missing IMDB IDs for Amazon Data

#### Goals

- Find missing titles from original merge.
- Find missing titles on IMDB using fuzzywuzzy.

**NOTE: We will use the same process as in 03.1_a_netflix_missing_title_ids notebook**

### 1. Import data

amazon = pd.read_csv("../Data/amazon_shows.csv")

amazon_ids = pd.read_pickle("../Data/amazon_ids.pkl")

amazon_missing = pd.read_pickle("../Data/amazon_missing.pkl")

### 2. Compare length of values of oringal list vs new ones

We will compare the number of rows from the original amazon data frame and the sum of the missing titles and the amazon ids.

len(amazon)

len(amazon_ids) + len(amazon_missing)

amazon["show"].nunique()

6 values are missing but when looking at unique values on the original data frame, we get the same result. This means that there must be some repeated titles which may be from different years. We will make a value count for the show column in the original amazon data frame.

amazon["show"].value_counts().head(6)

We will look at each of them and add them the amazon missing data frame.

amazon[amazon["show"] == "Moby Dick"]

amazon_ids[amazon_ids["show"] == "Moby Dick"]

amazon_missing = amazon_missing.append(amazon[(amazon["show"] == "Moby Dick") & (amazon["year"] == 2011)])

amazon[amazon["show"] == "Clifford the Big Red Dog"]

amazon_ids[amazon_ids["show"] == "Clifford the Big Red Dog"]

amazon_missing = amazon_missing.append(amazon[(amazon["show"] == "Clifford the Big Red Dog") & (amazon["year"] == 2019)])

amazon[amazon["show"] == "Undercover"]

amazon_ids[amazon_ids["show"] == "Undercover"]

amazon_missing = amazon_missing.append(amazon[(amazon["show"] == "Undercover") & (amazon["year"] == 2016)])

amazon[amazon["show"] == "Baby Talk"]

amazon_ids[amazon_ids["show"] == "Baby Talk"]

amazon_missing = amazon_missing.append(amazon[(amazon["show"] == "Baby Talk") & (amazon["year"] == 2017)])

amazon[amazon["show"] == "Elizabeth I"]

amazon_ids[amazon_ids["show"] == "Elizabeth I"]

amazon_missing = amazon_missing.append(amazon[(amazon["show"] == "Elizabeth I") & (amazon["year"] == 2017)])

amazon[amazon["show"] == "Rome"]

amazon_ids[amazon_ids["show"] == "Rome"]

amazon_missing = amazon_missing.append(amazon[(amazon["show"] == "Rome") & (amazon["year"] == 2012)])

len(amazon_missing) + len(amazon_ids)

Now we got the same values from the original tv shows and the new data frames.

amazon_missing = amazon_missing.reset_index(drop=True)

### 3. Find missing titles using fuzzywuzzy

We will first create a list for missing shows.

amazon_shows_missing = [show for show in amazon_missing["show"]]

len(amazon_shows_missing)

imdb = pd.read_pickle(path + "Data/imdb_tv_all.pkl")

imdb_titles = [title for title in imdb["primaryTitle"]]

We will use the same function as on the netflix notebook, this time we will look for values with a ration higher than 80, since we are looking at more missing shows, we need more tolerance in order to get correct data.

def find_shows(show):
    matches = []

    for title in imdb_titles:
        # compute ratio and remove case-sensitivity
        ratio = fuzz.ratio(title.lower(), show.lower())

        # add all matches to list with ratio > 80
        if ratio >= 80:
            matches.append((title, show, ratio))
    
    # return none if there was no match found
    if len(matches) == 0:
        return None
    return sorted(matches, key=lambda x: x[2], reverse=True)[0][0]

We will now create a new column with the matched imdb titles.

amazon_missing["imdb_titles"] = amazon_missing["show"].apply(lambda x: find_shows(x))

amazon_missing.head(10)

amazon_missing_2 = amazon_missing[amazon_missing["imdb_titles"].isna()].reset_index(drop=True)

amazon_missing = amazon_missing[~amazon_missing["imdb_titles"].isna()].reset_index(drop=True)

amazon_missing_2

We still have 324 titles that were not found. We will first merge the new imdb ids to the amazon_ids data frame and then continue to review the missing titles.

### 4. Merge IMDB IDs

We will still merge on the year, in order to be sure that we are not merging a remake or a similar title.

amazon_missing = amazon_missing.merge(imdb, how="left", left_on=["imdb_titles", "year"]
                                      , right_on=["primaryTitle", "startYear"])

We will separate the missing titles that were not matched.

amazon_missing_3 = amazon_missing[amazon_missing["tconst"].isna()].reset_index(drop=True)

amazon_missing = amazon_missing[~amazon_missing["tconst"].isna()].reset_index(drop=True)

We will join the new ids found to the amazon_ids dataframe.

amazon_ids = amazon_ids.append(amazon_missing).reset_index(drop=True)

amazon_missing_3 = amazon_missing_3[["show", "year", "rating", "imdb", "rotten_tomatoes", "imdb_titles"]]

We will take a look on the titles that weren't matched.

amazon_missing_3

Some of them don't have a matching title, we will now add a new column showing the fuzzy ratio and remove those below 90.

fuzzy_ratio = []

for row in range(len(amazon_missing_3)):
    fuzzy_ratio.append(fuzz.ratio(amazon_missing_3.loc[row, "show"].lower(), amazon_missing_3.loc[row, "imdb_titles"].lower()))

amazon_missing_3["fuzzy_ratio"] = fuzzy_ratio

amazon_missing_3_keep = amazon_missing_3[amazon_missing_3["fuzzy_ratio"] >= 90].drop(columns="fuzzy_ratio")

We will keep this titles and merge them with their IMDB ID, and add them to the amazon_ids data frame.

amazon_missing_3_keep = amazon_missing_3_keep.merge(imdb, how="left", left_on=["imdb_titles"], right_on=["primaryTitle"])

amazon_missing_3_keep[amazon_missing_3_keep["tconst"].isna()]

We don't have missing ids on this new data frame. We will now add it to the amazon_ids data frame.

len(amazon_missing_3)

len(amazon_missing_3_keep)

amazon_ids = amazon_ids.append(amazon_missing_3_keep).reset_index(drop=True)

amazon_ids = amazon_ids.drop_duplicates("tconst")

len(amazon_ids)/len(amazon)

We will now work on amazon_missing_2 data frame, we will look for imdb titles decreasing the ratio to 60.

def find_shows2(show):
    matches = []

    for title in imdb_titles:
        # compute ratio and remove case-sensitivity
        ratio = fuzz.ratio(title.lower(), show.lower())

        # add all matches to list with ratio > 60
        if ratio >= 60:
            matches.append((title, show, ratio))
    if len(matches) == 0:
        return None
    return sorted(matches, key=lambda x: x[2], reverse=True)[0][0]

amazon_missing_2["imdb_titles"] = amazon_missing_2["show"].apply(lambda x: find_shows2(x))

amazon_missing_2.head(20)

From the 20 first elements, it doesn't seem that we have found great matches. We will compute the fuzzy ratio in a new column and see if we can remove some.

fuzzy_ratio = []

for row in range(len(amazon_missing_2)):
    if amazon_missing_2.loc[row, "imdb_titles"]:
        fuzzy_ratio.append(fuzz.ratio(amazon_missing_2.loc[row, "show"].lower(), amazon_missing_2.loc[row, "imdb_titles"].lower()))
    else:
        fuzzy_ratio.append(None)
amazon_missing_2["fuzzy_ratio"] = fuzzy_ratio

amazon_missing_2[amazon_missing_2["fuzzy_ratio"] >= 75].head(20)

We can still see some titles that are not the same, so we will remove this values and work with what we have already.

We will now calculate the ratio of titles with an imdb id match and the original titles.

len(amazon_ids)/len(amazon)

This means that we will be using 75% of the original data.

### 5. Cleaning final data

We will create a final data frame including:
- show
- year
- rating
- imdb
- rotten_tomatoes
- imdb_id
- all data from title basics

First, we will need to create the imdb_id column, this will have tconst and titleId values.

amazon_ids["imdb_id"] = np.where(amazon_ids["tconst"].isna(), amazon_ids["titleId"], amazon_ids["tconst"])

amazon_ids["imdb_id"].isna().value_counts()

This means we have all the IMDB IDs for all the titles in this final data frame. We will now remove columns that are not show, year, rating, imdb, rotten_tomatoes and imdb_id and merge again.

to_drop = [col for col in amazon_ids.columns if col not in ["show", "year", "rating", "imdb", "rotten_tomatoes", "imdb_id"]]

amazon_ids = amazon_ids.drop(columns=to_drop)

amazon_ids.head()

Now we will merge all data from title basics.

amazon_ids = amazon_ids.merge(imdb, how="left", left_on="imdb_id", right_on="tconst")

amazon_ids.head()

amazon_ids = amazon_ids.drop(columns="tconst")

### 6. Export final data

# amazon_ids.to_pickle("../Data/amazon_final.pkl")
