==> Notebooks/03.2_c_hbo_missing_titles.ipynb <==
import pandas as pd
import numpy as np
from fuzzywuzzy import fuzz

## Find missing IMDB IDs for HBO Data

#### Goals

- Find missing titles from original merge.
- Find missing titles on IMDB using fuzzywuzzy.

**NOTE: We will use the same process as in 03.1_a_netflix_missing_title_ids notebook**

### 1. Import data

hbo = pd.read_csv("../Data/hbo_shows.csv")

hbo_ids = pd.read_pickle("../Data/hbo_ids.pkl")

hbo_missing = pd.read_pickle("../Data/hbo_missing.pkl")

### 2. Compare length of values of oringal list vs new ones

We will compare the number of rows from the original amazon data frame and the sum of the missing titles and the amazon ids.

len(hbo)

len(hbo_ids) + len(hbo_missing)

hbo["show"].nunique()

3 values are missing but when looking at unique values on the original data frame, we get the same result. This means that there must be some repeated titles which may be from different years. We will make a value count for the show column in the original hbo data frame.

hbo["show"].value_counts().head(3)

We will look at each of them and add them the amazon missing data frame.

hbo[hbo["show"] == "The Electric Company"]

hbo_ids[hbo_ids["show"] == "The Electric Company"]

hbo_missing = hbo_missing.append(hbo[(hbo["show"] == "The Electric Company") & (hbo["year"] == 2009)])

hbo[hbo["show"] == "High Maintenance"]

hbo_ids[hbo_ids["show"] == "High Maintenance"]

hbo_missing = hbo_missing.append(hbo[(hbo["show"] == "High Maintenance") & (hbo["year"] == 2012)])

hbo[hbo["show"] == "Rome"]

hbo_ids[hbo_ids["show"] == "Rome"]

hbo_missing = hbo_missing.append(hbo[(hbo["show"] == "Rome") & (hbo["year"] == 2012)])

len(hbo_missing) + len(hbo_ids)

Now we got the same values from the original tv shows and the new data frames.

hbo_missing = hbo_missing.reset_index(drop=True)

### 3. Find missing titles using fuzzywuzzy

We will first create a list for missing shows.

hbo_shows_missing = [show for show in hbo_missing["show"]]

len(hbo_shows_missing)

imdb = pd.read_pickle(path + "Data/imdb_tv_all.pkl")

imdb_titles = [title for title in imdb["primaryTitle"]]

We will use the same function as on the netflix notebook, this time we will look for values with a ration higher than 80, since for HBO it is easy to see this because we have a small data frame.

def find_shows(show):
    matches = []

    for title in imdb_titles:
        # compute ratio and remove case-sensitivity
        ratio = fuzz.ratio(title.lower(), show.lower())

        # add all matches to list with ratio > 80
        if ratio >= 80:
            matches.append((title, show, ratio))
    
    # return none if there was no match found
    if len(matches) == 0:
        return None
    return sorted(matches, key=lambda x: x[2], reverse=True)[0][0]

We will now create a new column with the matched imdb titles.

hbo_missing["imdb_titles"] = hbo_missing["show"].apply(lambda x: find_shows(x))

hbo_missing

We will remove the null values for imdb titles, this accounts for 2% of the data.


hbo_missing = hbo_missing[~hbo_missing["imdb_titles"].isna()].reset_index(drop=True)

### 4. Merge IMDB IDs

We will still merge on the year, in order to be sure that we are not merging a remake or a similar title.

hbo_missing = hbo_missing.merge(imdb, how="left", left_on=["imdb_titles", "year"], right_on=["primaryTitle", "startYear"])

We will separate the missing titles that were not matched.

hbo_missing_2 = hbo_missing[hbo_missing["tconst"].isna()].reset_index(drop=True)

hbo_missing_2

We will also remove this shows from the analysis since we cannot verify it is the same show due to the years, and some of them don't have a matching title.

hbo_missing = hbo_missing[~hbo_missing["tconst"].isna()]

hbo_missing = hbo_missing.reset_index(drop=True)

hbo_missing

hbo_ids = hbo_ids.append(hbo_missing).reset_index(drop=True).drop_duplicates("tconst")

hbo_ids

We will now calculate the ratio of titles with an imdb id match and the original titles.

len(hbo_ids)/len(hbo)

This means that we will be using 94.5% of the original data.

### 5. Cleaning final data

We will create a final data frame including:
- show
- year
- rating
- imdb
- rotten_tomatoes
- imdb_id
- all data from title basics

First, we will need to create the imdb_id column, this will have tconst and titleId values.

hbo_ids["imdb_id"] = np.where(hbo_ids["tconst"].isna(), hbo_ids["titleId"], hbo_ids["tconst"])

hbo_ids["imdb_id"].isna().value_counts()

This means we have all the IMDB IDs for all the titles in this final data frame. We will now remove columns that are not show, year, rating, imdb, rotten_tomatoes and imdb_id and merge again.

to_drop = [col for col in hbo_ids.columns if col not in ["show", "year", "rating", "imdb", "rotten_tomatoes", "imdb_id"]]

hbo_ids = hbo_ids.drop(columns=to_drop)

hbo_ids.head()

Now we will merge all data from title basics

hbo_ids = hbo_ids.merge(imdb, how="left", left_on="imdb_id", right_on="tconst")

hbo_ids.head()

### 6. Export final data

# hbo_ids.to_pickle("../Data/hbo_final.pkl")
