==> Notebooks/03.3_final_cleaning.ipynb <==
import pandas as pd
import numpy as np

pd.set_option("display.max_columns", None)

## Final Data Cleaning

#### Goals

- Check current data types
- Change data types if necessary
- Remove tv shows which years don't match with IMDB's year (We will remove this rows because we can not confirm if they are the same show, in order to have more accuracy on our analysis.)

----- 

## Netflix

### 1. Import data

netflix = pd.read_pickle("../Data/netflix_final.pkl")

netflix.head()

netflix.dtypes

### 2. Check null and unique values

len(netflix)

netflix.isna().sum()

netflix.nunique(axis=0)

From the previous cells we can see that:
- Most values are missing for rotten_tomatoes.
- isAdult has just one value.

We will drop this two columns, since we cannot get much information from it.


netflix = netflix.drop(columns=["rotten_tomatoes", "isAdult"])

### 3. Change data types

- rating: We will remove the '+' sign and turn it into an integer
- runtimeMinutes: we will change the type to integer

#### i. rating

netflix["rating"] = [str(i).replace("+", "") for i in netflix["rating"]]

netflix["rating"].value_counts()

Since rating is a string type, we will convert the 'nan' values to null and 'all' to 0, meaning that the series can be watched by all ages.

netflix["rating"] = np.where(netflix["rating"] == "nan", None, netflix["rating"])
netflix["rating"] = np.where(netflix["rating"] == "all", 0, netflix["rating"])

netflix["rating"].value_counts()

netflix["rating"] = pd.to_numeric(netflix["rating"], errors="coerce")

#### ii. runtimeMinutes

netflix["runtimeMinutes"] = pd.to_numeric(netflix["runtimeMinutes"], errors="coerce")

### 4. Rename rating and imdb columns

netflix = netflix.rename(columns={"rating":"age", "imdb":"imdb_rating"})

### 5. Check final data types

netflix.dtypes

### 6. Remove rows where year and startYear don't match and duplicates for imdb_id

netflix = netflix[netflix["year"] == netflix["startYear"]].reset_index(drop=True)

netflix.head()

netflix = netflix.drop_duplicates("imdb_id")

len(netflix) / 1915

We will be working with 85% from our original data.

### 7. Export data

# netflix.to_pickle("../Data/netflix_final_clean.pkl")

------

## Amazon

### 1. Import data

amazon = pd.read_pickle("../Data/amazon_final.pkl")

amazon.head()

amazon.dtypes

### 2. Check null and unique values

len(amazon)

amazon.isna().sum()

amazon.nunique(axis=0)

amazon["isAdult"].value_counts()

Since there is just one movie with one true value for isAdult, we will drop this column as well and follow the same process as for netflix.

amazon = amazon.drop(columns=["rotten_tomatoes", "isAdult"])

### 3. Change data types


amazon["rating"] = [str(i).replace("+", "") for i in amazon["rating"]]

amazon["rating"].value_counts()

amazon["rating"] = np.where(amazon["rating"] == "nan", None, amazon["rating"])
amazon["rating"] = np.where(amazon["rating"] == "all", 0, amazon["rating"])

amazon["rating"] = pd.to_numeric(amazon["rating"], errors="coerce")

amazon["runtimeMinutes"] = pd.to_numeric(amazon["runtimeMinutes"], errors="coerce")

### 4. Rename rating and imdb columns

amazon = amazon.rename(columns={"rating":"age", "imdb":"imdb_rating"})

### 5. Check final data types

amazon.dtypes

### 6. Remove rows where year and startYear don't match and duplicates for imdb_id

amazon = amazon[amazon["year"] == amazon["startYear"]].reset_index(drop=True)

amazon.head()

amazon = amazon.drop_duplicates("imdb_id")

len(amazon) / 2136

We will be working with 60% from our original data.

### 7. Export data


# amazon.to_pickle("../Data/amazon_final_clean.pkl")

------

## HBO

### 1. Import data

hbo = pd.read_pickle("../Data/hbo_final.pkl")

hbo.head()

hbo.dtypes

### 2. Check null and unique values

len(hbo)

hbo.isna().sum()

hbo.nunique(axis=0)

We will follow the same process as with Netflix.

hbo = hbo.drop(columns=["rotten_tomatoes", "isAdult"])

### 3. Change data types

hbo["rating"] = [str(i).replace("+", "") for i in hbo["rating"]]

hbo["rating"].value_counts()

hbo["rating"] = np.where(hbo["rating"] == "nan", None, hbo["rating"])
hbo["rating"] = np.where(hbo["rating"] == "all", 0, hbo["rating"])

hbo["rating"] = pd.to_numeric(hbo["rating"], errors="coerce")

hbo["runtimeMinutes"] = pd.to_numeric(hbo["runtimeMinutes"], errors="coerce")

### 4. Rename rating and imdb columns

hbo = hbo.rename(columns={"rating":"age", "imdb":"imdb_rating"})

### 5. Check final data types

hbo.dtypes

### 6. Remove rows where year and startYear don't match and duplicates for imdb_id

hbo = hbo[hbo["year"] == hbo["startYear"]].reset_index(drop=True)

hbo.head()

hbo = hbo.drop_duplicates("imdb_id")

len(hbo) / 200

We will be working with 86% from our original data.

### 7. Export data

# hbo.to_pickle("../Data/hbo_final_clean.pkl")
