==> Notebooks/06_streaming_service_recommender_df.ipynb <==
import pandas as pd
import numpy as np

pd.set_option("display.max_columns", None)

path = "C:/Users/Admin/Documents/ironhack/streaming_service_recommender/"

## Streaming Service Recommender
#### Goals

- Build data frame to be used for recommender

We want our final data frame to llok like the following:

| streaming_service | Drama | Comedy | Animation | ... |
|-------------------|-------|--------|-----------|-----|
| **Netflix**       | 0.3765 | 0.2735 | 0.1680   | ....|
| **Amazon**        | 0.3349 | 0.2432 | 0.1619   | ... |
| ... | ... | ... |... | ...|

Our genres should be our columns and the streaming services should be the index.

### 1. Import data

netflix = pd.read_pickle(path + "Data/netflix_genres.pkl")

amazon = pd.read_pickle(path + "Data/amazon_genres.pkl")

hbo = pd.read_pickle(path + "Data/hbo_genres.pkl")

### 2. Get list of total genres

We will first get the lists of genres for each streaming service, we will them add them together and afterwards convert it to a set to remove duplicates.

netflix_genres = netflix["genre"].to_list()

amazon_genres = amazon["genre"].to_list()

hbo_genres = hbo["genre"].to_list()

len(netflix_genres)

total_genres = list(set(netflix_genres + amazon_genres + hbo_genres))

len(total_genres)

total_genres.sort()

total_genres

Looks like the three streaming services shared the same genres because we have the same number of genres as for Netflix and Amazon, HBO has 24.

### 3. Create new data frame

Since they share the same genres, we will prepare each data frame, then append them with each other and finally creating a pivot table which we will use for the recommender.

#### i. Prepare each data frame

netflix["streaming_service"] = "Netflix"

netflix_genres = netflix[["genre", "ratio", "streaming_service"]]

netflix_genres.head()

amazon["streaming_service"] = "Amazon"

amazon_genres = amazon[["genre", "ratio", "streaming_service"]]

amazon_genres.head()

hbo["streaming_service"] = "HBO"

hbo_genres = hbo[["genre", "ratio", "streaming_service"]]

hbo_genres.head()

#### ii. Append data frames

genres_features = netflix_genres.append(amazon_genres).append(hbo_genres)

#### iii. Create pivot table

genres_features = netflix_genres.append(amazon_genres).append(hbo_genres).reset_index()

genres_recommender = genres_features.pivot_table(index="streaming_service",
                                      columns="genre",
                                      values="ratio").fillna(0)

genres_recommender

### 4. Export final data frame

# genres_recommender.to_pickle(path + "Data/genres_recommender.pkl")
