==> Notebooks/06.2_user_df.ipynb <==
import pandas as pd
import numpy as np
import re
from fuzzywuzzy import fuzz

pd.set_option("display.max_columns", None)

# User Data Frame

#### Goals

- Create function to read newest input survey and transform it to the same format as genres_recommender in order to append it and get the most similar streaming service.

Survey Link: https://forms.gle/Jb7gzK7VMbnvpRvP9

### 1. Import data

url = "https://docs.google.com/spreadsheets/d/1VsL_x7WAcEm-pLNmCqNwrDlFjTR_PUDpiadyUYJM-AQ/export?format=csv&gid=1360685521"

survey = pd.read_csv(url)

We will remove the first 7 columns since they were used for testing the survey.

survey = survey.loc[7:].reset_index(drop=True)

survey

Rows 1-3 were used to test picking only Netflix shows (first row), then Amazon shows (second row) and finally HBO shows (third row). We will start working with the first row, in order to test Netflix and later on we will create a function for Amazon and HBO using the same process.

We will also need to join all streaming services genres data frames in order to obtain the genres data frame for our survey users.

netflix = pd.read_pickle("../Data/netflix_genres.pkl")
amazon = pd.read_pickle("../Data/amazon_genres.pkl")
hbo = pd.read_pickle("../Data/hbo_genres.pkl")

all_shows = netflix.append(amazon).append(hbo).reset_index(drop=True)

all_shows.to_pickle("../Data/all_shows_genres.pkl")

### 2. Remove Timestamp column

We will remove the Timestamp column since it would not be useful for our analysis.

survey = survey.drop(columns="Timestamp")

### 3. Get list of TV shows for survey user

all_shows_lst = [show for show in all_shows["show"]]

# using regex, we will get the title of TV shows selected

titles_netflix = []

for show in survey.loc[0].values:
    # remove the year and description
    title_pattern = r"(.*)\s\(\d{4}\).*$"
    title = re.findall(title_pattern, show)
    titles_netflix.append(title[0])


matches_netflix = []

for title in titles_netflix:
    for show in all_shows_lst:
        
        # compute ratio and remove case-sensitivity
        ratio = fuzz.ratio(title.lower(), show.lower())

        # add titles to list with ratio = 100
        if ratio == 100:
            matches_netflix.append(show)
            
            # we will remove duplicates since there might be shared TV shows between streaming services
            matches_netflix = list(set(matches_netflix))

matches_netflix

It works, now we will get the data frame for the shows listed on matches_netflix.

### 4. Create data frame to use for recommender model

user_netflix = all_shows[all_shows["show"].isin(matches_netflix)].reset_index(drop=True)

user_netflix

We will follow the same process as done on 05_genres and 06.1_streaming_service_recommender_df

genre_columns = [col for col in user_netflix.columns if col not in ["show", "genres"]]

user_netflix_ratio = (pd.DataFrame(user_netflix[genre_columns].sum()
                                    .sort_values(ascending=False))
                       .reset_index()
                       .rename(columns={"index":"genre", 0:"frequency"}))

user_netflix_ratio["ratio"] = round((user_netflix_ratio["frequency"] / 30), 4)

user_netflix_ratio["user"] = "User"

user_netflix_recommender = user_netflix_ratio.pivot_table(index="user",
                                      columns="genre",
                                      values="ratio")

user_netflix_recommender

Now we have our final dataframe, we will now create a function that sums up all of this steps.

### 5. Create function

# the input of this function will be the df survey, all_shows df, genre_columns and row location

def get_recommender_df(survey_df, df_all_shows, genre_columns, row):
    
    survey_df = survey_df.drop(columns="Timestamp")
    
    all_shows_lst = [show for show in df_all_shows["show"]]
    
    titles = []

    for show in survey_df.loc[row].values:
        # remove the year and description
        title_pattern = r"(.*)\s\(\d{4}\).*$"
        title = re.findall(title_pattern, show)
        titles.append(title[0])

    matches = []

    for title in titles:
        for show in all_shows_lst:
            ratio = fuzz.ratio(title.lower(), show.lower())
            if ratio == 100:
                matches.append(show)
                matches = list(set(matches))
                              
    user_df = df_all_shows[df_all_shows["show"].isin(matches)].reset_index(drop=True)
    
    user_genre_ratio = (pd.DataFrame(user_df[genre_columns].sum()
                                    .sort_values(ascending=False))
                       .reset_index()
                       .rename(columns={"index":"genre", 0:"frequency"}))

    user_genre_ratio["ratio"] = round((user_genre_ratio["frequency"] / 30), 4)

    user_genre_ratio["user"] = "User"

    user_recommender = user_genre_ratio.pivot_table(index="user",
                                          columns="genre",
                                          values="ratio")
    
    return user_recommender

In order to test the function, we will import everything needed again.

url = "https://docs.google.com/spreadsheets/d/1VsL_x7WAcEm-pLNmCqNwrDlFjTR_PUDpiadyUYJM-AQ/export?format=csv&gid=1360685521"

survey_test = pd.read_csv(url)
survey_test = survey_test.loc[7:].reset_index(drop=True)

all_shows_test = pd.read_pickle("../Data/all_shows_genres.pkl")

genre_columns = [col for col in all_shows_test.columns if col not in ["show", "genres"]]

user_amazon_recommender = get_recommender_df(survey_test, all_shows_test, genre_columns, 1)

user_amazon_recommender

user_hbo_recommender = get_recommender_df(survey_test, all_shows_test, genre_columns, 2)

user_hbo_recommender

It works, now we will export our data to test our model.

### 6. Export data

# user_netflix_recommender.to_pickle("../Data/netflix_test.pkl")

# user_amazon_recommender.to_pickle("../Data/amazon_test.pkl")

# user_hbo_recommender.to_pickle("../Data/hbo_test.pkl")

### 7. Update function

For our recommender, we will need always the last row from our survey data frame, so we will update the function in order to get just the last column. We will also have the genre_columns inside the function.

def get_recommender_df(survey_df, df_all_shows):
    
    genre_columns = ['Crime', 'Drama', 'Thriller', 'Fantasy', 'Horror', 'Mystery', 'Comedy', 'Sci-Fi', 'Biography',
                 'Action', 'Adventure', 'Romance', 'History', 'Documentary', 'Animation', 'War', 'Sport',
                 'Family', 'Western', 'Short', 'Reality-TV', 'Musical', 'Music', 'Game-Show', 'Talk-Show', 'News']
    
    survey_df = survey_df.drop(columns="Timestamp")
    
    all_shows_lst = [show for show in df_all_shows["show"]]
    
    titles = []
        
    # we will select just the last element using the tail pandas method
    for show in survey_df.tail(1).values[0]:
        # remove the year and description
        title_pattern = r"(.*)\s\(\d{4}\).*$"
        title = re.findall(title_pattern, show)
        titles.append(title[0])

    matches = []

    for title in titles:
        for show in all_shows_lst:
            ratio = fuzz.ratio(title.lower(), show.lower())
            if ratio == 100:
                matches.append(show)
                matches = list(set(matches))
                              
    user_df = df_all_shows[df_all_shows["show"].isin(matches)].reset_index(drop=True)
    
    user_genre_ratio = (pd.DataFrame(user_df[genre_columns].sum()
                                    .sort_values(ascending=False))
                       .reset_index()
                       .rename(columns={"index":"genre", 0:"frequency"}))

    user_genre_ratio["ratio"] = round((user_genre_ratio["frequency"] / 30), 4)

    user_genre_ratio["user"] = "User"

    user_recommender = user_genre_ratio.pivot_table(index="user",
                                          columns="genre",
                                          values="ratio")
    
    return user_recommender
